Move use-filters into `higher-order` folder
